Moved VGA driver to hwpnp tree. Fixed misc display bugs.

+++ src/modules/hwpnp/vga/textmode_ops.cpp
@@ -125,7 +125,7 @@ uint8_t make_color(uint8_t fg, uint8_t bg)
 
 uint16_t make_vgaentry(char c, uint8_t color)
 {
-    uint16_t c16 = c;
+    uint16_t c16 = (uint8_t)c;
     uint16_t color16 = color;
     return c16 | color16 << 8;
 }
